Data Validation
Data validation should be complete in so the logic works. Might eventually need to move this validation to the day file as I am having problems thinking of how to represent the keys in the map. Maybe just the month and day are the actual key and the rest of the string gets passed to a validation method in Day class?